Added Binary_search.c file.
I have added the file Binary_search.c .  It contains complete implementation of the binary search algorithm in c language.
I have added comments for almost every step which makes it beginner friendly and it could be understood by anyone. I have also explained and defined the binary search algorithm in comments at the top of the program. It is helpful for everyone including students and those preparing for placements. It is helpful for both theory as i have added all the necessary theory for the binary search algorithm and implementation as it can easily be implemented by anyone.